testing/integration-tests/requirements.txt
# YAWL Integration Tests Requirements

# Testing Framework
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-timeout>=2.1.0
pytest-xdist>=2.5.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# API Testing
requests>=2.27.0
urllib3>=1.26.0

# Database Testing
psycopg2-binary>=2.9.0

# Docker Testing
docker>=5.0.0

# Kubernetes Testing
kubernetes>=20.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=0.19.0

# Reporting
jinja2>=3.0.0
